+++ mobile/android/fenix/app/src/main/java/org/mozilla/fenix/components/Core.kt
@@ -178,6 +178,7 @@ class Core(
                 store.state.desktopMode
             },
             webContentIsolationStrategy = WebContentIsolationStrategy.ISOLATE_HIGH_VALUE,
+            fetchPriorityEnabled = FxNimbus.features.networking.value().fetchPriorityEnabled,
         )
 
         // Apply fingerprinting protection overrides if the feature is enabled in Nimbus